moved analytics function to analytics_clr.py, import methods
